Fixed another sprite bug and expanded map

+++ DungeonCrawler/Renderer.cpp
@@ -22,7 +22,7 @@ Renderer::Renderer() : viewPort(HEIGHT - PANEL_HEIGHT, WIDTH)
 }
 
 void Renderer::Draw(Game &game) {
-    memcpy(pixels, viewPort.Pixels(), viewPort.Width() * viewPort.Height() * sizeof(unsigned int));
+    memcpy(pixels, viewPort.Pixels(), static_cast<size_t>(viewPort.Width() * viewPort.Height() * sizeof(int)));
     //Note UI frame doesn't need to be redrawn
     viewPort.Draw(game);
 }
